+++ src/Caliburn.Light.Core/ViewModel/IConductor.cs
@@ -9,7 +9,7 @@ namespace Caliburn.Light;
 /// </summary>
 /// <remarks>
 /// Conducted instances can opt-in to the life-cycle by implementing any of the following
-/// <see cref="IActivatable"/>, <see cref="ICloseGuard"/>, <see cref="IChild"/>.
+/// <see cref="IActivatable"/>, <see cref="ICloseGuard"/>, <see cref="IParentAware"/>.
 /// </remarks>
 public interface IConductor
 {